[DD-159] feature: Directory Hierarchy Refactoring
