+++ Cargo.toml
@@ -3,6 +3,13 @@ name = "myco_client_rust"
 version = "0.1.0"
 edition = "2018"
 
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "myco_client_rust"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
