+++ src/algorithms/mod.rs
@@ -1,0 +1,2 @@
+mod pay_as_bid;
+pub use pay_as_bid::PayAsBid;
