+++ src/connectors/mod.rs
@@ -1,0 +1,2 @@
+mod redis_subscriber;
+pub use redis_subscriber::psubscribe;
